app/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
torch
faiss-gpu
beautifulsoup4
requests
pandas
numpy
pydantic
